Refactor: Header component structure
