+++ src/pages/Home/index.tsx
@@ -1,6 +1,6 @@
+import { Header } from '../../components';
 import Footer from "../../components/Footer/Footer";
-import Header from "../../components/Header/Header";
-import { Wellcome, Challenges, Contact, Projects, AboutMe, Skills } from '../../sections/'
+import { Wellcome, Challenges, Contact, Projects, AboutMe, Skills } from '../../sections/';
 
 export const Home = () => {
   return (
